% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_alpha.R
\docType{class}
\name{alpha-class}
\alias{alpha-class}
\title{Alpha Class}
\description{
This class represents an extension of the \code{easynem} class.
}
\section{Slots}{

\describe{
\item{\code{result}}{The calculation results of storage alpha diversity.}
}}

