% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacotu.R
\docType{data}
\name{bacotu}
\alias{bacotu}
\title{Bacterial OTUs sample data}
\format{
A data frame with 8695 rows and 33 variables:
\describe{
\item{OTUID}{Operational Taxonomic Units code}
\item{Y11_Spring_1}{The first processing of microbial data in spring of 11 consecutive cropping years. Other columns are interpreted similarly.}
}
}
\source{
\url{luqiaofang1991@163.com}
}
\usage{
data(bacotu)
}
\description{
Bacterial OTUs data were recorded in different seasons in different years of continuous cropping.
}
\examples{
data(bacotu)
head(bacotu)
}
\keyword{datasets}
