% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bactax.R
\docType{data}
\name{bactax}
\alias{bactax}
\title{Sample data on bacterial classification}
\format{
A data frame with 8695 rows and 8 variables:
\describe{
\item{OTUID}{Operational Taxonomic Units code}
\item{Kingdom}{The Kingdom level classification of different microbial OTUs. Other columns are interpreted similarly.}
}
}
\source{
\url{luqiaofang1991@163.com}
}
\usage{
data(bactax)
}
\description{
The 7-level classification system of different bacterial OTUs.
}
\examples{
data(bactax)
head(bactax)
}
\keyword{datasets}
