% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_beta.R
\docType{class}
\name{beta-class}
\alias{beta-class}
\title{An S4 class to store beta diversity results.}
\description{
An S4 class to store beta diversity results.
}
\section{Slots}{

\describe{
\item{\code{meta}}{A data frame of meta data.}

\item{\code{result}}{A data frame of pairwise comparison results.}

\item{\code{temp}}{A character vector of the difference comparison.}
}}

