% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{beta2-class}
\alias{beta2-class}
\title{A class to store beta diversity results (two-factor)}
\description{
\code{beta2-class} is used to store the results of beta diversity analysis,
including results for drawing and comparing differences between groups.
}
\details{
Users can construct a \code{beta2-class} through \code{\link{calc_beta2}},
which can then be connected to \code{\link{nem_plot}} to visualize the results.
}
\section{Slots}{

\describe{
\item{\code{meta}}{A data frame storing basic elements for visualization.}

\item{\code{result}}{A character of pairwise comparison results.}

\item{\code{temp}}{A character vector of the difference comparison.}
}}

\seealso{
The constructor, \code{\link{calc_beta2}}; Class for storing single factor beta
diversity analysis, \code{\link{beta-class}}; Visualization function,
\code{\link{nem_plot}}.
}
