% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_alpha.R
\name{calc_alpha}
\alias{calc_alpha}
\title{calc_alpha}
\usage{
calc_alpha(data, ...)
}
\arguments{
\item{data}{easynem type data.}

\item{...}{Other parameters for calc_alpha.}
}
\value{
An alpha object.
}
\description{
Alpha diversity calculation for the tab.
}
