% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_compare.R
\name{calc_compare}
\alias{calc_compare}
\title{calc_compare}
\usage{
calc_compare(data, .group, y, method, ...)
}
\arguments{
\item{data}{easynem type data.}

\item{.group}{The group variable.}

\item{y}{The dependent variable.}

\item{method}{The method of difference comparison.}

\item{...}{Other parameters for method functions.}
}
\value{
An compare object.
}
\description{
Differences between groups were compared.
}
