% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_compare2.R
\name{calc_compare2}
\alias{calc_compare2}
\title{calc_compare2}
\usage{
calc_compare2(data, .group1, .group2, y, method, ...)
}
\arguments{
\item{data}{easynem type data.}

\item{.group1}{The group variable.}

\item{.group2}{The group variable.}

\item{y}{The dependent variable.}

\item{method}{The method of difference comparison.}

\item{...}{Other parameters for method functions.}
}
\value{
An compare2 object.
}
\description{
Differences between groups were compared.
}
