% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_funguild2.R
\name{calc_funguild2}
\alias{calc_funguild2}
\title{calc_funguild2}
\usage{
calc_funguild2(data, .group1, .group2)
}
\arguments{
\item{data}{nemindex type data.}

\item{.group1}{The group variable.}

\item{.group2}{The group variable.}
}
\value{
An funguild2 object.
}
\description{
Calculate funguild2 between treatments.
}
