% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mf2.R
\name{calc_mf2}
\alias{calc_mf2}
\title{calc_mf2}
\usage{
calc_mf2(data, .group1, .group2)
}
\arguments{
\item{data}{nemindex type data.}

\item{.group1}{The group variable.}

\item{.group2}{The group variable.}
}
\value{
An mf2 object.
}
\description{
Calculate Metabolic footprints between treatments.
}
