% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ter.R
\name{calc_ter}
\alias{calc_ter}
\title{calc_ter}
\usage{
calc_ter(data, .group)
}
\arguments{
\item{data}{easynem type data.}

\item{.group}{The group variable.}
}
\value{
An ter object.
}
\description{
Calculate ternary analysis between treatments.
}
