% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ter2.R
\name{calc_ter2}
\alias{calc_ter2}
\title{calc_ter2}
\usage{
calc_ter2(data, .group1, .group2)
}
\arguments{
\item{data}{easynem type data.}

\item{.group1}{The group variable.}

\item{.group2}{The group variable.}
}
\value{
An ter2 object.
}
\description{
Calculate ternary analysis between treatments.
}
