% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTest2.R
\docType{class}
\name{compare2-class}
\alias{compare2-class}
\title{An S4 class to store multiple comparisons results.}
\description{
An S4 class to store multiple comparisons results.
}
\section{Slots}{

\describe{
\item{\code{meta}}{A data frame of meta data.}

\item{\code{result}}{A data frame of multiple comparisons results.}

\item{\code{temp}}{A character vector of the difference comparison.}
}}

