% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nem_plot.R
\name{compare2}
\alias{compare2}
\alias{nem_plot,compare2-method}
\title{nem_plot}
\usage{
\S4method{nem_plot}{compare2}(object, type1 = 1, type2 = 1, add, ...)
}
\arguments{
\item{object}{easynem or other types data.}

\item{type1}{Drawing style, bar or box.}

\item{type2}{Drawing style, tufted or faceted.}

\item{add}{Add mean and standard error or mean and standard deviation.}

\item{...}{Other parameters for nem_plot functions.}
}
\value{
An ggplot object.
}
\description{
For visualization of nematode community data.
}
