% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\docType{data}
\name{design}
\alias{design}
\title{Sample data from the experimental design table}
\format{
A data frame with 32 rows and 4 variables:
\describe{
\item{SampleID}{Sample number.}
\item{con_crop}{Soil continuous cropping years.}
\item{season}{Soil sampling season.}
\item{condition}{Results of soil nematode disease status.}
}
}
\source{
\url{luqiaofang1991@163.com}
}
\usage{
data(design)
}
\description{
The banana rhizosphere soil was collected in spring and summer for 2,5,8,11 years of continuous cropping.
}
\examples{
data(design)
head(design)
}
\keyword{datasets}
