% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemdatabase.R
\docType{data}
\name{.nemdatabase}
\alias{.nemdatabase}
\title{Nematode database query}
\format{
An object of class \code{shiny.appobj} of length 5.
}
\usage{
.nemdatabase
}
\description{
Nematode database query
}
\keyword{datasets}
