% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merge.R
\docType{class}
\name{easynem-class}
\alias{easynem-class}
\title{An S4 class to represent microbial community structure data set.}
\description{
An S4 class to represent microbial community structure data set.
}
\section{Slots}{

\describe{
\item{\code{tab}}{Otutab.}

\item{\code{tax}}{Species classification table.}

\item{\code{design}}{Experimental design table.}

\item{\code{output}}{Other result data table.}

\item{\code{temp}}{Temporary data file.}

\item{\code{env}}{Environmental factors.}
}}

