% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easynem-package.R
\docType{package}
\name{easynem-package}
\alias{easynem}
\alias{easynem-package}
\title{easynem: An R package for joint analysis of nematode community indices}
\description{
This R package currently only supports analysis and calculation related to nematode community ecological indices, but will introduce microbial community analysis and various biological community analysis functions as well as soil biological interaction analysis functions in the future. This R package will be maintained and updated continuously.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/whkygl/easynem}
  \item Report bugs at \url{https://github.com/whkygl/easynem/issues}
}

}
\author{
\strong{Maintainer}: Wang Kunguang \email{whkygl@163.com} (\href{https://orcid.org/0000-0001-7384-5002}{ORCID})

}
\keyword{internal}
