% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easynem_example.R
\name{easynem_example}
\alias{easynem_example}
\title{Path to example files}
\usage{
easynem_example(path = NULL)
}
\arguments{
\item{path}{The path to the example files.}
}
\value{
The path to the example files.
}
\description{
This function returns the path to the example files.
}
