% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ef.R
\docType{class}
\name{ef-class}
\alias{ef-class}
\title{An S4 class to store energy flow results.}
\description{
An S4 class to store energy flow results.
}
\section{Slots}{

\describe{
\item{\code{result}}{A data frame of energy flow results.}
}}

