% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nem_plot.R
\name{ef}
\alias{ef}
\alias{nem_plot,ef-method}
\title{nem_plot}
\usage{
\S4method{nem_plot}{ef}(object)
}
\arguments{
\item{object}{ef or other types data.}
}
\value{
An ggplot object.
}
\description{
For visualization of nematode community data.
}
