% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nem_plot.R
\name{ef2}
\alias{ef2}
\alias{nem_plot,ef2-method}
\title{nem_plot}
\usage{
\S4method{nem_plot}{ef2}(object)
}
\arguments{
\item{object}{ef2 or other types data.}
}
\value{
An ggplot object.
}
\description{
For visualization of nematode community data.
}
