% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\docType{data}
\name{env}
\alias{env}
\title{Soil environmental factor representation data}
\format{
A data frame with 32 rows and 13 variables:
\describe{
\item{SampleID}{Sample number.}
\item{pH}{Soil pH index. Other columns are interpreted similarly.}
}
}
\source{
\url{luqiaofang1991@163.com}
}
\usage{
data(env)
}
\description{
Environmental factors such as soil physical and chemical properties were recorded during different continuous cropping years.
}
\examples{
data(env)
head(env)
}
\keyword{datasets}
