% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemdatabase.R
\name{filter_data}
\alias{filter_data}
\title{Nematode database query}
\usage{
filter_data(df, keywords)
}
\arguments{
\item{df}{Intermediate variable}

\item{keywords}{Intermediate variable
.nemdatabase}

\item{filter_data}{Intermediate variable}
}
\description{
Nematode database query
}
