% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_name.R
\name{filter_name}
\alias{filter_name}
\title{filter_name}
\usage{
filter_name(data, target, ...)
}
\arguments{
\item{data}{easynem type data.}

\item{target}{tab, tax or meta, where tab represents the species abundance table, tax represents the species classification table, and meta represents the metadata table.}

\item{...}{Other parameters of the filter function of the dplyr package.}
}
\value{
An easynem object.
}
\description{
Used to filter easynem type data by name.
}
