% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_funguild.R
\docType{class}
\name{funguild-class}
\alias{funguild-class}
\title{An S4 class to store funguild results.}
\description{
An S4 class to store funguild results.
}
\section{Slots}{

\describe{
\item{\code{result}}{A data frame of funguild results.}
}}

