% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nem_plot.R
\name{funguild}
\alias{funguild}
\alias{nem_plot,funguild-method}
\title{nem_plot}
\usage{
\S4method{nem_plot}{funguild}(object, ...)
}
\arguments{
\item{object}{funguild or other types data.}

\item{...}{Other parameters for ggplot2 functions.}
}
\value{
An ggplot object.
}
\description{
For visualization of nematode community data.
}
