% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_funguild2.R
\docType{class}
\name{funguild2-class}
\alias{funguild2-class}
\title{An S4 class to store funguild2 results.}
\description{
An S4 class to store funguild2 results.
}
\section{Slots}{

\describe{
\item{\code{result}}{A data frame of funguild2 results.}
}}

