% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nem_plot.R
\name{funguild2}
\alias{funguild2}
\alias{nem_plot,funguild2-method}
\title{nem_plot}
\usage{
\S4method{nem_plot}{funguild2}(object, ...)
}
\arguments{
\item{object}{funguild2 or other types data.}

\item{...}{Other parameters for ggplot2 functions.}
}
\value{
An ggplot object.
}
\description{
For visualization of nematode community data.
}
