% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funotu.R
\docType{data}
\name{funotu}
\alias{funotu}
\title{Fungal OTUs sample data}
\format{
A data frame with 717 rows and 33 variables:
\describe{
\item{OTUID}{Operational Taxonomic Units code}
\item{Y11_Spring_1}{The first processing of microbial data in spring of 11 consecutive cropping years. Other columns are interpreted similarly.}
}
}
\source{
\url{luqiaofang1991@163.com}
}
\usage{
data(funotu)
}
\description{
Fungal OTUs data were recorded in different seasons in different years of continuous cropping.
}
\examples{
data(funotu)
head(funotu)
}
\keyword{datasets}
