% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSD.R
\name{HSD}
\alias{HSD}
\title{HSD test for grouped data.}
\usage{
HSD(data, .group, y, ...)
}
\arguments{
\item{data}{easynem type data.}

\item{.group}{The group variable.}

\item{y}{Dependent variable.}

\item{...}{Other parameters for TukeyHSD.}
}
\value{
An compare object.
}
\description{
Multiple comparisons of treatments by means of HSD and a grouping of treatments. The level by alpha default is 0.05. Returns p-values adjusted using one of several methods.
}
