% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSD2.R
\name{HSD2}
\alias{HSD2}
\title{HSD2}
\usage{
HSD2(data, .group1, .group2, y, ...)
}
\arguments{
\item{data}{easynem type data.}

\item{.group1}{The group variable.}

\item{.group2}{The group variable.}

\item{y}{Dependent variable.}

\item{...}{Other parameters for HSD2.}
}
\value{
An compare2 object.
}
\description{
HSD test was performed for grouped data.
}
