% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{lme-class}
\alias{lme-class}
\title{A S4 class to store the linear regression analysis results (single factor)}
\description{
The \code{lme-class} is used to store the results of linear regression analysis.
}
\section{Slots}{

\describe{
\item{\code{meta}}{Stores the data frame for plotting.}

\item{\code{result}}{A data frame for storing the results of linear regression analysis.}
}}

\seealso{
The constructor, \code{\link{calc_lm}}; Visualization function, \code{\link{nem_plot}}.
}
