% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSD2.R
\name{LSD2}
\alias{LSD2}
\title{LSD2}
\usage{
LSD2(data, .group1, .group2, y, method2 = "none", ...)
}
\arguments{
\item{data}{easynem type data.}

\item{.group1}{The group variable.}

\item{.group2}{The group variable.}

\item{y}{Dependent variable.}

\item{method2}{The method for adjust p value.}

\item{...}{Other parameters for LSD2.}
}
\value{
An compare2 object.
}
\description{
LSD test was performed for grouped data.
}
