% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{mf-class}
\alias{mf-class}
\title{A S4 class to store Metabolic footprints results (single factor)}
\description{
The \code{mf-class} is used to store the results of nematode metabolic footprints analysis.
}
\section{Slots}{

\describe{
\item{\code{result}}{A data frame for storing the results of metabolic footprinting analysis.}
}}

\seealso{
The constructor, \code{\link{calc_mf}}; Visualization function, \code{\link{nem_plot}}.
}
