% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mf2.R
\docType{class}
\name{mf2-class}
\alias{mf2-class}
\title{An S4 class to store Metabolic footprints results.}
\description{
An S4 class to store Metabolic footprints results.
}
\section{Slots}{

\describe{
\item{\code{result}}{A data frame of Metabolic footprints results.}
}}

