% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nem_plot.R
\name{mf2}
\alias{mf2}
\alias{nem_plot,mf2-method}
\title{nem_plot}
\usage{
\S4method{nem_plot}{mf2}(object, kei = 1, ksi = 1, ...)
}
\arguments{
\item{object}{mf or other types data.}

\item{kei}{Adjust the proportion of the vertical axis of the graph.}

\item{ksi}{Adjust the proportion of the abscissa of the graph.}

\item{...}{Other parameters for ggplot2 functions.}
}
\value{
An ggplot object.
}
\description{
For visualization of nematode community data.
}
