% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merge.R
\docType{class}
\name{nem-class}
\alias{nem-class}
\title{An S4 class to represent nematode community structure data set.}
\description{
An S4 class to represent nematode community structure data set.
}
\section{Slots}{

\describe{
\item{\code{tab}}{Otutab.}

\item{\code{tax}}{Species classification table.}

\item{\code{design}}{Experimental design table.}

\item{\code{output}}{Other result data table.}

\item{\code{temp}}{Temporary data file.}

\item{\code{env}}{Environmental factors.}

\item{\code{nemdb}}{Nematode function index reference database.}
}}

