% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nem_calc.R
\name{nem_calc}
\alias{nem_calc}
\title{calc}
\usage{
nem_calc(data, f, ...)
}
\arguments{
\item{data}{easynem type data.}

\item{f}{Function parameters for microbial or nematode community calculations.}

\item{...}{Other parameters.}
}
\value{
easynem or other data types.
}
\description{
For microbial or nematode community calculations.
}
