% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nem_database.R
\name{nem_database}
\alias{nem_database}
\title{Nematode database search}
\usage{
nem_database()
}
\value{
A web interface
}
\description{
This function provides a visual interface for retrieving basic data of nematodes.
The database used is from \url{http://nemaplex.ucdavis.edu/Ecology/EcophysiologyParms/EcoParameterMenu.html}
}
