% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nem_trans.R
\name{nem_trans}
\alias{nem_trans}
\title{nem_trans}
\usage{
nem_trans(data, f, ...)
}
\arguments{
\item{data}{easynem type data.}

\item{f}{Function parameters for data filtering and transformation.}

\item{...}{Other parameters.}
}
\value{
An easynem object.
}
\description{
Used to convert and filter easynem type data.
}
