% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nemindex.R
\docType{class}
\name{nemindex-class}
\alias{nemindex-class}
\title{nemindex Class}
\description{
This class represents an extension of the \code{easynem} class.
}
\section{Slots}{

\describe{
\item{\code{result}}{The calculation results of storage nemindex.}
}}

