% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemotu.R
\docType{data}
\name{nemotu}
\alias{nemotu}
\title{Nematode OTUs sample data}
\format{
A data frame with 145 rows and 33 variables:
\describe{
\item{OTUID}{Operational Taxonomic Units code}
\item{Y11_Spring_1}{The first processing of microbial data in spring of 11 consecutive cropping years. Other columns are interpreted similarly.}
}
}
\source{
\url{luqiaofang1991@163.com}
}
\usage{
data(nemotu)
}
\description{
Nematode OTUs data were recorded in different seasons in different years of continuous cropping.
}
\examples{
data(nemotu)
head(nemotu)
}
\keyword{datasets}
