% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_factor.R
\name{order_factor}
\alias{order_factor}
\title{order_factor}
\usage{
order_factor(data, group, order)
}
\arguments{
\item{data}{easynem type data.}

\item{group}{Selection of meta columns.}

\item{order}{Order of factors.}
}
\value{
An easynem object.
}
\description{
Meta attributes of easynem grouping factors in order to rearrangement.
}
