% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merge.R
\name{read_design,easynem,character-method}
\alias{read_design,easynem,character-method}
\title{A read_design method written for easynem}
\usage{
\S4method{read_design}{easynem,character}(object, design, ...)
}
\arguments{
\item{object}{easynem.}

\item{design}{character.}

\item{...}{Readr package read_csv function other parameters.}
}
\value{
easynem.
}
\description{
A read_design method written for easynem
}
