% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merge.R
\name{read_design,nem,character-method}
\alias{read_design,nem,character-method}
\title{A read_design method written for nem}
\usage{
\S4method{read_design}{nem,character}(object, design, ...)
}
\arguments{
\item{object}{nem.}

\item{design}{character.}

\item{...}{Readr package read_csv function other parameters.}
}
\value{
nem.
}
\description{
A read_design method written for nem
}
