% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merge.R
\name{read_design}
\alias{read_design}
\title{read_design: Generic functions related to reading into experimental design table}
\usage{
read_design(object, design, ...)
}
\arguments{
\item{object}{Microbial data type: easynem or nem.}

\item{design}{The string associated with the file name.}

\item{...}{Readr package read_csv function other parameters.}
}
\value{
Returns a microbiome dataset of type easynem containing properties such as tab, tax, design, env, etc.
}
\description{
read_design: Generic functions related to reading into experimental design table
}
