% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merge.R
\name{read_env,easynem,character-method}
\alias{read_env,easynem,character-method}
\title{A read_env method written for easynem}
\usage{
\S4method{read_env}{easynem,character}(object, env, ...)
}
\arguments{
\item{object}{easynem.}

\item{env}{character.}

\item{...}{Readr package read_csv function other parameters.}
}
\value{
easynem.
}
\description{
A read_env method written for easynem
}
