% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merge.R
\name{read_env,nem,character-method}
\alias{read_env,nem,character-method}
\title{A read_env method written for nem}
\usage{
\S4method{read_env}{nem,character}(object, env, ...)
}
\arguments{
\item{object}{nem.}

\item{env}{character.}

\item{...}{Readr package read_csv function other parameters.}
}
\value{
nem.
}
\description{
A read_env method written for nem
}
