% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merge.R
\name{read_env}
\alias{read_env}
\title{read_env: Generic functions related to reading into table of environmental factors}
\usage{
read_env(object, env, ...)
}
\arguments{
\item{object}{Microbial data type: easynem or nem.}

\item{env}{The string associated with the file name.}

\item{...}{Readr package read_csv function other parameters}
}
\value{
Returns a microbiome dataset of type easynem containing properties such as tab, tax, design, env, etc.
}
\description{
read_env: Generic functions related to reading into table of environmental factors
}
