% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nem.R
\name{read_nem}
\alias{read_nem}
\title{read_nem}
\usage{
read_nem(tab = 0, tax = 0, meta = 0, ...)
}
\arguments{
\item{tab}{Otutab table.}

\item{tax}{Taxonomy table.}

\item{meta}{Metadata table.}

\item{...}{Other default parameters for read_csv function.}
}
\value{
An easynem object.
}
\description{
Used for reading Otutab, Taxonomy, and Metadata tables.
Currently, this function only supports text documents in csv format.
}
\examples{
bac <- read_nem(tab = easynem_example("bacotu.csv"), tax = easynem_example("bactax.csv"))
show(bac)
}
