% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nem2.R
\name{read_nem2}
\alias{read_nem2}
\title{read_nem2}
\usage{
read_nem2(tab = 0, tax = 0, meta = 0, ...)
}
\arguments{
\item{tab}{Otutab table.}

\item{tax}{Taxonomy table.}

\item{meta}{Metadata table.}

\item{...}{Other default parameters for read_nem2 function.}
}
\value{
An easynem object.
}
\description{
Used for reading Otutab, Taxonomy, and Metadata tables.
Currently, this function only supports text documents in dataframe format.
}
