% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merge.R
\name{read_tab,easynem,character-method}
\alias{read_tab,easynem,character-method}
\title{A read_tab method written for easynem}
\usage{
\S4method{read_tab}{easynem,character}(object, tab, ...)
}
\arguments{
\item{object}{easynem.}

\item{tab}{character.}

\item{...}{Readr package read_csv function other parameters.}
}
\value{
easynem.
}
\description{
A read_tab method written for easynem
}
