% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merge.R
\name{read_tab,nem,character-method}
\alias{read_tab,nem,character-method}
\title{A read_tab method written for nem}
\usage{
\S4method{read_tab}{nem,character}(object, tab, ...)
}
\arguments{
\item{object}{nem.}

\item{tab}{character.}

\item{...}{Readr package read_csv function other parameters.}
}
\value{
nem.
}
\description{
A read_tab method written for nem
}
