% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merge.R
\name{read_tab}
\alias{read_tab}
\title{read_tab: Generic functions related to reading into otutab}
\usage{
read_tab(object, tab, ...)
}
\arguments{
\item{object}{Microbial data type: easynem or nem.}

\item{tab}{The string associated with the file name.}

\item{...}{Readr package read_csv function other parameters}
}
\value{
Returns a microbiome dataset of type easynem containing properties such as tab, tax, design, env, etc.
}
\description{
read_tab: Generic functions related to reading into otutab
}
