% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merge.R
\name{read_tax,easynem,character-method}
\alias{read_tax,easynem,character-method}
\title{A read_tax method written for easynem}
\usage{
\S4method{read_tax}{easynem,character}(object, tax, ...)
}
\arguments{
\item{object}{easynem.}

\item{tax}{character.}

\item{...}{Readr package read_csv function other parameters.}
}
\value{
easynem.
}
\description{
A read_tax method written for easynem
}
