% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merge.R
\name{read_tax,nem,character-method}
\alias{read_tax,nem,character-method}
\title{A read_tax method written for nem}
\usage{
\S4method{read_tax}{nem,character}(object, tax, ...)
}
\arguments{
\item{object}{nem.}

\item{tax}{character.}

\item{...}{Readr package read_csv function other parameters.}
}
\value{
nem.
}
\description{
A read_tax method written for nem
}
