% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merge.R
\name{read_tax}
\alias{read_tax}
\title{read_tax: Generic functions related to reading into species classification table}
\usage{
read_tax(object, tax, ...)
}
\arguments{
\item{object}{Microbial data type: easynem or nem.}

\item{tax}{The string associated with the file name.}

\item{...}{Readr package read_csv function other parameters}
}
\value{
Returns a microbiome dataset of type easynem containing properties such as tab, tax, design, env, etc.
}
\description{
read_tax: Generic functions related to reading into species classification table
}
