% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merge.R
\name{show,easynem-method}
\alias{show,easynem-method}
\title{Easynem's show general function}
\usage{
\S4method{show}{easynem}(object)
}
\arguments{
\item{object}{Microbial community data set.}
}
\value{
Print out the easynem object in the console window.
}
\description{
Easynem's show general function
}
