% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merge.R
\name{show,nem-method}
\alias{show,nem-method}
\title{Nem's show method}
\usage{
\S4method{show}{nem}(object)
}
\arguments{
\item{object}{Nematode community structure data set.}
}
\value{
Print out the nem object details in the console window.
}
\description{
Nem's show method
}
