% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merge.R
\name{summary,easynem-method}
\alias{summary,easynem-method}
\title{Easynem's summary method}
\usage{
\S4method{summary}{easynem}(object)
}
\arguments{
\item{object}{Microbial data type: easynem.}
}
\value{
Print out the easynem object details in the console window.
}
\description{
Easynem's summary method
}
