% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merge.R
\name{summary}
\alias{summary}
\title{Easynem's summary general function}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{Microbial data type: easynem or nem.}

\item{...}{Others.}
}
\value{
Print out the easynem object details in the console window.
}
\description{
Easynem's summary general function
}
