% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ter-class}
\alias{ter-class}
\title{A S4 class to store the ternary analysis results (single factor)}
\description{
The \code{ter-class} is used to store the results of nematode ternary analysis.
}
\section{Slots}{

\describe{
\item{\code{result}}{A data frame for storing the results of ternary analysis.}
}}

\seealso{
The constructor, \code{\link{calc_ter}}; Visualization function, \code{\link{nem_plot}}.
}
