% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nem_plot.R
\name{ter}
\alias{ter}
\alias{nem_plot,ter-method}
\title{nem_plot}
\usage{
\S4method{nem_plot}{ter}(object, type)
}
\arguments{
\item{object}{ter or other types data.}

\item{type}{Select the ternary graph type, cp or feeding.}
}
\value{
An ggplot object.
}
\description{
For visualization of nematode community data.
}
