% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_formula.R
\name{trans_formula}
\alias{trans_formula}
\title{trans_formula}
\usage{
trans_formula(data, var, formu, ...)
}
\arguments{
\item{data}{easynem type data.}

\item{var}{Variable name to be converted.}

\item{formu}{Formula parameters for data conversion.}

\item{...}{Other parameters of the vegan package decostand function.}
}
\value{
An easynem object.
}
\description{
Convert data according to a formula.
}
