% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_name.R
\name{trans_name}
\alias{trans_name}
\title{trans_name}
\usage{
trans_name(data, taxonomy, ...)
}
\arguments{
\item{data}{easynem type data.}

\item{taxonomy}{Species classification name.}

\item{...}{Other parameters.}
}
\value{
An easynem object.
}
\description{
Used to sum species abundance tables grouped by taxonomy table.
}
