% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTest2.R
\name{TTest2}
\alias{TTest2}
\title{TTest2}
\usage{
TTest2(data, .group1, .group2, y, ...)
}
\arguments{
\item{data}{easynem type data.}

\item{.group1}{The group variable.}

\item{.group2}{The group variable.}

\item{y}{Dependent variable.}

\item{...}{Other parameters for TTest2.}
}
\value{
An compare2 object.
}
\description{
t test was performed for grouped data.
}
