% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{alpha-class}
\alias{alpha-class}
\title{Class for storing alpha diversity calculation results}
\description{
The \code{alpha-class} is an extension of the \code{\link{easynem-class}} to store
the results of alpha diversity calculations.
}
\section{Slots}{

\describe{
\item{\code{result}}{The calculation results of storage alpha diversity.}
}}

\seealso{
The constructor, \code{\link{calc_alpha}}; Visualization function, \code{\link{nem_plot}}.
}
