% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_beta.R
\name{calc_beta}
\alias{calc_beta}
\title{calc_beta}
\usage{
calc_beta(data, type, .group, method, ...)
}
\arguments{
\item{data}{easynem type data.}

\item{type}{pca, pcoa or nmds.}

\item{.group}{The group variable.}

\item{method}{The method of calculating the distance matrix.}

\item{...}{Other parameters for cmdscale and vegdist functions.}
}
\value{
An beta object.
}
\description{
Calculate beata diversity and generate results for multiple comparisons between treatments.
}
