% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_beta2.R
\name{calc_beta2}
\alias{calc_beta2}
\title{calc_beta2}
\usage{
calc_beta2(data, type, .group1, .group2, method, ...)
}
\arguments{
\item{data}{easynem type data.}

\item{type}{pca, pcoa or nmds.}

\item{.group1}{The group variable.}

\item{.group2}{The group variable.}

\item{method}{The method of calculating the distance matrix.}

\item{...}{Other parameters for cmdscale and vegdist functions.}
}
\value{
An beta2 object.
}
\description{
Calculate beata diversity and generate results for multiple comparisons between treatments.
}
