% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ef.R
\name{calc_ef}
\alias{calc_ef}
\title{calc_ef}
\usage{
calc_ef(data, .group)
}
\arguments{
\item{data}{nemindex type data.}

\item{.group}{The group variable.}
}
\value{
An ef object.
}
\description{
Calculate energy flow between treatments.
}
