% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ef2.R
\name{calc_ef2}
\alias{calc_ef2}
\title{calc_ef2}
\usage{
calc_ef2(data, .group1, .group2)
}
\arguments{
\item{data}{nemindex type data.}

\item{.group1}{The group variable.}

\item{.group2}{The group variable.}
}
\value{
An ef2 object.
}
\description{
Calculate energy flow between treatments.
}
