% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_funguild.R
\name{calc_funguild}
\alias{calc_funguild}
\title{calc_funguild}
\usage{
calc_funguild(data, .group)
}
\arguments{
\item{data}{nemindex type data.}

\item{.group}{The group variable.}
}
\value{
An funguild object.
}
\description{
Calculate calc_funguild between treatments.
}
