% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mf.R
\name{calc_mf}
\alias{calc_mf}
\title{calc_mf}
\usage{
calc_mf(data, .group)
}
\arguments{
\item{data}{nemindex type data.}

\item{.group}{The group variable.}
}
\value{
An mf object.
}
\description{
Calculate Metabolic footprints between treatments.
}
