% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nemindex.R
\name{calc_nemindex}
\alias{calc_nemindex}
\title{calc_nemindex}
\usage{
calc_nemindex(data, ...)
}
\arguments{
\item{data}{easynem type data.}

\item{...}{Other parameters for calc_nemindex.}
}
\value{
An nemindex object.
}
\description{
nematode index calculation for the tab.
}
