% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{compare-class}
\alias{compare-class}
\title{A S4 class to store multiple comparisons results (single factor).}
\description{
\code{compare-class} is used to store the results of multiple comparisons results,
including results for drawing and comparing differences between groups.
}
\details{
Users can construct a \code{compare-class} through \code{\link{calc_compare}},
which can then be connected to \code{\link{nem_plot}} to visualize the results.
}
\section{Slots}{

\describe{
\item{\code{meta}}{A data frame storing basic elements for visualization.}

\item{\code{result}}{A data frame of multiple comparisons results.}

\item{\code{temp}}{A character vector of the difference comparison.}
}}

\seealso{
The constructor, \code{\link{calc_compare}}; Class for storing two-factor
multiple comparisons analysis, \code{\link{compare2-class}}; Visualization function,
\code{\link{nem_plot}}.
}
