% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nem_plot.R
\name{compare}
\alias{compare}
\alias{nem_plot,compare-method}
\title{nem_plot}
\usage{
\S4method{nem_plot}{compare}(object, type = 1, ...)
}
\arguments{
\item{object}{easynem or other types data.}

\item{type}{Drawing style, bar or box.}

\item{...}{Other parameters for nem_plot functions.}
}
\value{
An ggplot object.
}
\description{
For visualization of nematode community data.
}
