% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{compare2-class}
\alias{compare2-class}
\title{An S4 class to store multiple comparisons results (two-factor).}
\description{
\code{compare2-class} is used to store the results of multiple comparisons results,
including results for drawing and comparing differences between groups.
}
\details{
Users can construct a \code{compare2-class} through \code{\link{calc_compare2}},
which can then be connected to \code{\link{nem_plot}} to visualize the results.
}
\section{Slots}{

\describe{
\item{\code{meta}}{A data frame storing basic elements for visualization.}

\item{\code{result}}{A data frame of multiple comparisons results.}

\item{\code{temp}}{A character vector of the difference comparison.}
}}

\seealso{
The constructor, \code{\link{calc_compare2}}; Class for storing single factor
multiple comparisons analysis, \code{\link{compare-class}}; Visualization function,
\code{\link{nem_plot}}.
}
