% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nem.R
\docType{class}
\name{easynem-class}
\alias{easynem-class}
\title{easynem Class}
\description{
The \code{easynem} class represents a base class that includes two slots, \code{slot1} and \code{slot2}.
This class can be extended by other classes to provide additional functionality.
}
\section{Slots}{

\describe{
\item{\code{tab}}{A numeric value representing the first slot. Provide a detailed description of what this slot is used for.}

\item{\code{tax}}{A character string representing the second slot. Provide a detailed description of what this slot is used for.}

\item{\code{meta}}{Storage experiment design and result.}
}}

