% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{easynem-class}
\alias{easynem-class}
\title{The main experiment-level class for easynem data}
\description{
Integrate the nematode abundance table, nematode classification table, and
experimental design table into an \code{easynem-class}, which makes it easier to
filter and manage nematode data, and easier to link to the nematode database
and conduct subsequent analysis.
}
\details{
Users can read data via \code{\link{read_nem}} or \code{\link{read_nem2}}.
When there are missing slots in easynem, the system will issue a warning, but
this will not affect subsequent analysis.
}
\section{Slots}{

\describe{
\item{\code{tab}}{A single object of nematode abundance table.}

\item{\code{tax}}{A single object of nematode classification table.}

\item{\code{meta}}{A single object of experimental design table.}
}}

\seealso{
The constructor, \code{\link{read_nem}} for reading csv files and
\code{\link{read_nem2}} for reading tibble type data.
}
