% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easynem-package.R
\docType{package}
\name{easynem-package}
\alias{easynem}
\alias{easynem-package}
\title{easynem: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/whkygl/easynem}
  \item Report bugs at \url{https://github.com/whkygl/easynem/issues}
}

}
\author{
\strong{Maintainer}: Wang Kunguang \email{whkygl@163.com} (\href{https://orcid.org/0000-0001-7384-5002}{ORCID})

}
\keyword{internal}
