% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ef-class}
\alias{ef-class}
\title{A S4 class to store energy flow results (single factor)}
\description{
The \code{ef-class} is an extension of the \code{\link{easynem-class}} to store
the results of nematode energy flow analysis.
}
\section{Slots}{

\describe{
\item{\code{result}}{A data frame for storing the results of energy flow analysis.}
}}

\seealso{
The constructor, \code{\link{calc_ef}}; Visualization function, \code{\link{nem_plot}}.
}
