% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\docType{data}
\name{env}
\alias{env}
\title{Soil environmental factor representation data}
\format{
A data frame with 32 rows and 13 variables:
\describe{
\item{SampleID}{Sample number.}
\item{pH}{Soil pH index.}
\item{SOC}{Soil SOC content.}
\item{NH4}{Soil NH4 content.}
\item{NO3}{Soil NO3 content.}
\item{P}{Soil P content.}
\item{K}{Soil K content.}
\item{Ca}{Soil Ca content.}
\item{Mg}{Soil Mg content.}
\item{Fe}{Soil Fe content.}
\item{Mn}{Soil Mn content.}
\item{Cu}{Soil Cu content.}
\item{Zn}{Soil Zn content.}
}
}
\source{
\url{luqiaofang1991@163.com}
}
\usage{
data(env)
}
\description{
Environmental factors such as soil physical and chemical properties were recorded during different continuous cropping years.
}
\examples{
data(env)
head(env)
}
\keyword{datasets}
