% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_num.R
\name{filter_num}
\alias{filter_num}
\title{filter_num}
\usage{
filter_num(data, num, ...)
}
\arguments{
\item{data}{easynem type data.}

\item{num}{Filter threshold value.}

\item{...}{Other parameters.}
}
\value{
An easynem object.
}
\description{
Used to filter species whose quantity or discovery rate is higher than num. If num is greater than 1, filter by quantity; if num is less than 1, filter by discovery rate.
}
