% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{funguild2-class}
\alias{funguild2-class}
\title{Class for storing computational results of nematode functional guild analysis (two-factor)}
\description{
The \code{funguild2-class} is used to store the results of nematode functional guild analysis.
}
\section{Slots}{

\describe{
\item{\code{result}}{A data frame of storing computational results of nematode
functional guild analysis.}
}}

\seealso{
The constructor, \code{\link{calc_funguild2}}; Visualization function, \code{\link{nem_plot}}.
}
