% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funotu.R
\docType{data}
\name{funotu}
\alias{funotu}
\title{Fungal OTUs sample data}
\format{
A data frame with 717 rows and 33 variables:
\describe{
\item{OTUID}{Operational Taxonomic Units code}
\item{Y11_Spring_1}{The first processing of microbial data in spring of 11 consecutive cropping years.}
\item{Y11_Spring_2}{The second processing of microbial data in spring of 11 consecutive cropping years.}
\item{Y11_Spring_3}{The third processing of microbial data in spring of 11 consecutive cropping years.}
\item{Y11_Spring_4}{The forth processing of microbial data in spring of 11 consecutive cropping years.}
\item{Y11_Summer_1}{The first processing of microbial data in summer of 11 consecutive cropping years.}
\item{Y11_Summer_2}{The second processing of microbial data in summer of 11 consecutive cropping years.}
\item{Y11_Summer_3}{The third processing of microbial data in summer of 11 consecutive cropping years.}
\item{Y11_Summer_4}{The forth processing of microbial data in summer of 11 consecutive cropping years.}
\item{Y2_Spring_1}{The first processing of microbial data in spring of 2 consecutive cropping years.}
\item{Y2_Spring_2}{The second processing of microbial data in spring of 2 consecutive cropping years.}
\item{Y2_Spring_3}{The third processing of microbial data in spring of 2 consecutive cropping years.}
\item{Y2_Spring_4}{The forth processing of microbial data in spring of 2 consecutive cropping years.}
\item{Y2_Summer_1}{The first processing of microbial data in summer of 2 consecutive cropping years.}
\item{Y2_Summer_2}{The second processing of microbial data in summer of 2 consecutive cropping years.}
\item{Y2_Summer_3}{The third processing of microbial data in summer of 2 consecutive cropping years.}
\item{Y2_Summer_4}{The forth processing of microbial data in summer of 2 consecutive cropping years.}
\item{Y5_Spring_1}{The first processing of microbial data in spring of 5 consecutive cropping years.}
\item{Y5_Spring_2}{The second processing of microbial data in spring of 5 consecutive cropping years.}
\item{Y5_Spring_3}{The third processing of microbial data in spring of 5 consecutive cropping years.}
\item{Y5_Spring_4}{The forth processing of microbial data in spring of 5 consecutive cropping years.}
\item{Y5_Summer_1}{The first processing of microbial data in summer of 5 consecutive cropping years.}
\item{Y5_Summer_2}{The second processing of microbial data in summer of 5 consecutive cropping years.}
\item{Y5_Summer_3}{The third processing of microbial data in summer of 5 consecutive cropping years.}
\item{Y5_Summer_4}{The forth processing of microbial data in summer of 5 consecutive cropping years.}
\item{Y8_Spring_1}{The first processing of microbial data in spring of 8 consecutive cropping years.}
\item{Y8_Spring_2}{The second processing of microbial data in spring of 8 consecutive cropping years.}
\item{Y8_Spring_3}{The third processing of microbial data in spring of 8 consecutive cropping years.}
\item{Y8_Spring_4}{The forth processing of microbial data in spring of 8 consecutive cropping years.}
\item{Y8_Summer_1}{The first processing of microbial data in summer of 8 consecutive cropping years.}
\item{Y8_Summer_2}{The second processing of microbial data in summer of 8 consecutive cropping years.}
\item{Y8_Summer_3}{The third processing of microbial data in summer of 8 consecutive cropping years.}
\item{Y8_Summer_4}{The forth processing of microbial data in summer of 8 consecutive cropping years.}
}
}
\source{
\url{luqiaofang1991@163.com}
}
\usage{
data(funotu)
}
\description{
Fungal OTUs data were recorded in different seasons in different years of continuous cropping.
}
\examples{
data(funotu)
head(funotu)
}
\keyword{datasets}
