% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funtax.R
\docType{data}
\name{funtax}
\alias{funtax}
\title{Sample data on fungal classification}
\format{
A data frame with 717 rows and 8 variables:
\describe{
\item{OTUID}{Operational Taxonomic Units code}
\item{Kingdom}{The Kingdom level classification of different microbial OTUs.}
\item{Phylum}{The Phylum level classification of different microbial OTUs.}
\item{Class}{The Class level classification of different microbial OTUs.}
\item{Order}{The Order level classification of different microbial OTUs.}
\item{Family}{The Family level classification of different microbial OTUs.}
\item{Genus}{The Genus level classification of different microbial OTUs.}
\item{Species}{The Species level classification of different microbial OTUs.}
}
}
\source{
\url{luqiaofang1991@163.com}
}
\usage{
data(funtax)
}
\description{
The 7-level classification system of different fungal OTUs.
}
\examples{
data(funtax)
head(funtax)
}
\keyword{datasets}
