% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KruskalTest.R
\name{KruskalTest}
\alias{KruskalTest}
\title{KruskalTest}
\usage{
KruskalTest(
  data,
  .group,
  y,
  exact = FALSE,
  sort = TRUE,
  .method = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"),
  ...
)
}
\arguments{
\item{data}{easynem type data.}

\item{.group}{The group variable.}

\item{y}{Dependent variable.}

\item{exact}{a logical indicating whether an exact p-value should be computed.}

\item{sort}{logical. If sort groups by median dependent variable values.}

\item{.method}{method for correcting p-values for multiple comparisons.}

\item{...}{Other parameters for KruskalTest.}
}
\value{
An compare object.
}
\description{
Kruskal-Wallis test was performed for grouped data.
}
