% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KruskalTest2.R
\name{KruskalTest2}
\alias{KruskalTest2}
\title{KruskalTest2}
\usage{
KruskalTest2(data, .group1, .group2, y, method2 = "none", ...)
}
\arguments{
\item{data}{easynem type data.}

\item{.group1}{The group variable.}

\item{.group2}{The group variable.}

\item{y}{Dependent variable.}

\item{method2}{The method for adjust p value.}

\item{...}{Other parameters for KruskalTest2.}
}
\value{
An compare2 object.
}
\description{
Kruskal-Wallis test was performed for grouped data.
}
