% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nem_database.R
\name{nem_database}
\alias{nem_database}
\title{nem_database}
\usage{
nem_database()
}
\value{
A web interface
}
\description{
nem_database is a function generated based on the shiny package for nematode retrieval.
}
