% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nem_plot.R
\name{nem_plot}
\alias{nem_plot}
\title{nem_plot}
\usage{
nem_plot(object, ...)
}
\arguments{
\item{object}{easynem or other types data.}

\item{...}{Other parameters.}
}
\description{
This is a generic function for visualization.
}
