% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{nemindex-class}
\alias{nemindex-class}
\title{Class for storing nematode ecological index calculation results}
\description{
The \code{nemindex-class} is an extension of the \code{\link{easynem-class}} to store
the results of nematode ecological index calculations.
}
\section{Slots}{

\describe{
\item{\code{result}}{The calculation results of storage nematode ecological index.}
}}

\seealso{
The constructor, \code{\link{calc_nemindex}}; Visualization function, \code{\link{nem_plot}}.
}
