% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemtax.R
\docType{data}
\name{nemtax}
\alias{nemtax}
\title{Sample data on nematode classification}
\format{
A data frame with 145 rows and 8 variables:
\describe{
\item{OTUID}{Operational Taxonomic Units code}
\item{Kingdom}{The Kingdom level classification of different microbial OTUs.}
\item{Phylum}{The Phylum level classification of different microbial OTUs.}
\item{Class}{The Class level classification of different microbial OTUs.}
\item{Order}{The Order level classification of different microbial OTUs.}
\item{Family}{The Family level classification of different microbial OTUs.}
\item{Genus}{The Genus level classification of different microbial OTUs.}
\item{Species}{The Species level classification of different microbial OTUs.}
}
}
\source{
\url{luqiaofang1991@163.com}
}
\usage{
data(nemtax)
}
\description{
The 7-level classification system of different nematode OTUs.
}
\examples{
data(nemtax)
head(nemtax)
}
\keyword{datasets}
