% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ter2.R
\docType{class}
\name{ter2-class}
\alias{ter2-class}
\title{An S4 class to store the ternary analysis results.}
\description{
An S4 class to store the ternary analysis results.
}
\section{Slots}{

\describe{
\item{\code{result}}{A data frame of the ternary analysis results.}
}}

