% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_combine.R
\name{trans_combine}
\alias{trans_combine}
\title{trans_formula}
\usage{
trans_combine(data, col)
}
\arguments{
\item{data}{easynem type data.}

\item{col}{The name of the column to be summed.}
}
\value{
An easynem object.
}
\description{
Sum meta by column.
}
