% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_formula_v.R
\name{trans_formula_v}
\alias{trans_formula_v}
\title{trans_formula_v}
\usage{
trans_formula_v(data, var, formu)
}
\arguments{
\item{data}{easynem type data.}

\item{var}{Variable name to be converted.}

\item{formu}{Formula parameters for data conversion.}
}
\value{
An easynem object.
}
\description{
Convert data according to a formula.
}
