% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_norm.R
\name{trans_norm}
\alias{trans_norm}
\title{trans_norm}
\usage{
trans_norm(data, method, MARGIN = 2, ...)
}
\arguments{
\item{data}{easynem type data.}

\item{method}{Methods used in standardization.}

\item{MARGIN}{The dimension to standardize. 1 for rows, 2 for columns.}

\item{...}{Other parameters of the vegan package decostand function.}
}
\value{
An easynem object.
}
\description{
Standardize species numbers to percentages or other.
}
