% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_rare.R
\name{trans_rare}
\alias{trans_rare}
\title{trans_rare}
\usage{
trans_rare(data, sample = 0, ...)
}
\arguments{
\item{data}{easynem type data.}

\item{sample}{The number of occurrences in the sample. If sample = 0, the minimum value is used by default.}

\item{...}{Other parameters of the vegan package rrarefy function.}
}
\value{
An easynem object.
}
\description{
Used to rare the species abundance table according to the number of occurrences in the sample. The default is to rare according to the minimum value.
}
