% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTest.R
\name{TTest}
\alias{TTest}
\title{TTest}
\usage{
TTest(data, .group, y, ...)
}
\arguments{
\item{data}{easynem type data.}

\item{.group}{The group variable.}

\item{y}{Dependent variable.}

\item{...}{Other parameters for TTest.}
}
\value{
An compare object.
}
\description{
t test was performed for grouped data.
}
