% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WilcoxTest.R
\name{WilcoxTest}
\alias{WilcoxTest}
\title{WilcoxTest}
\usage{
WilcoxTest(data, .group, y, ...)
}
\arguments{
\item{data}{easynem type data.}

\item{.group}{The group variable.}

\item{y}{Dependent variable.}

\item{...}{Other parameters for WilcoxTest.}
}
\value{
An compare object.
}
\description{
Wilcoxon test was performed for grouped data.
}
