% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WilcoxTest2.R
\name{WilcoxTest2}
\alias{WilcoxTest2}
\title{WilcoxTest2}
\usage{
WilcoxTest2(data, .group1, .group2, y, ...)
}
\arguments{
\item{data}{easynem type data.}

\item{.group1}{The group variable.}

\item{.group2}{The group variable.}

\item{y}{Dependent variable.}

\item{...}{Other parameters for WilcoxTest2.}
}
\value{
An compare2 object.
}
\description{
t test was performed for grouped data.
}
